Main Script
This is the main script that will run the assigned test. In this script there are references to outside functions which are also included in the
GitHub project folder. The purpose of this test is to login to a list of user’s cable boxes and search for specific movie titles, artwork, metadata etc. for each of the
assigned titles and return data for processing and reporting. This is fully automated and can be scheduled to automatically run at any time eliminating the need for
overhead which was the primary purpose of this project.