Function
This is a comprehensive error check. In the event navigation is interrupted this script will identify that it has and attempt to navigate back to the
correct location on screen. This will also check to make sure the remote image is still visible to the text reader and if not, attempt to correct that issue. This will also
correct itself if the user’s account crashes.