Function
This is a repeat script designed to navigate down a very long list of titles alphabetically. The script will read all the text in a designated
rectangle on the screen looking for each title in the pre-assigned list. Once a title has been located the script will record all the necessary data for that title and
proceed down the list. This script will know where it is in the alphabet while it’s navigating and if it determines it has gone beyond the last title in the search list, it
will end based on the LastLetter Script.